+++ pytation/test/tmodule.py
@@ -1,4 +1,4 @@
-# Copyright 2018-2021 Jetperch LLC
+# Copyright 2019-2021 Jetperch LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,6 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from . import analyze, cli, gui
 
-__all__ = ['analyze', 'cli', 'gui']
+def analyze(context):
+    pass
+
+
+def run(context, config=None):
+    return 0, {'hello':'world'}